Refactor BuildHandler into view_build.py.
