+++ gubernator/main.py
@@ -15,19 +15,15 @@
 # limitations under the License.
 
 import json
-import logging
-import re
 import os
 
 import webapp2
 
-import defusedxml.ElementTree as ET
-
 import github.models as ghm
 import gcs_async
-import log_parser
 import pull_request
-from view_base import RenderingHandler, memcache_memoize, gcs_ls
+import view_build
+from view_base import RenderingHandler, memcache_memoize, pad_numbers, PR_PREFIX
 import view_logs
 
 DEFAULT_JOBS = {
@@ -45,83 +41,6 @@ DEFAULT_JOBS = {
     }
 }
 
-PR_PREFIX = 'kubernetes-jenkins/pr-logs/pull'
-
-
-def pad_numbers(s):
-    """Modify a string to make its numbers suitable for natural sorting."""
-    return re.sub(r'\d+', lambda m: m.group(0).rjust(16, '0'), s)
-
-
-def parse_junit(xml, filename):
-    """Generate failed tests as a series of (name, duration, text, filename) tuples."""
-    tree = ET.fromstring(xml)
-    if tree.tag == 'testsuite':
-        for child in tree:
-            name = child.attrib['name']
-            time = float(child.attrib['time'])
-            for param in child.findall('failure'):
-                yield name, time, param.text, filename
-    elif tree.tag == 'testsuites':
-        for testsuite in tree:
-            suite_name = testsuite.attrib['name']
-            for child in testsuite.findall('testcase'):
-                name = '%s %s' % (suite_name, child.attrib['name'])
-                time = float(child.attrib['time'])
-                for param in child.findall('failure'):
-                    yield name, time, param.text, filename
-    else:
-        logging.error('unable to find failures, unexpected tag %s', tree.tag)
-
-
-@memcache_memoize('build-details://', expires=60 * 60 * 4)
-def build_details(build_dir):
-    """
-    Collect information from a build directory.
-
-    Args:
-        build_dir: GCS path containing a build's results.
-    Returns:
-        started: value from started.json {'version': ..., 'timestamp': ...}
-        finished: value from finished.json {'timestamp': ..., 'result': ...}
-        failures: list of (name, duration, text) tuples
-        build_log: a hilighted portion of errors in the build log. May be None.
-    """
-    started_fut = gcs_async.read(build_dir + '/started.json')
-    finished = gcs_async.read(build_dir + '/finished.json').get_result()
-    started = started_fut.get_result()
-    if finished and not started:
-        started = 'null'
-    if started and not finished:
-        finished = 'null'
-    elif not (started and finished):
-        return
-    started = json.loads(started)
-    finished = json.loads(finished)
-
-    failures = []
-    junit_paths = [f.filename for f in gcs_ls('%s/artifacts' % build_dir)
-                   if re.match(r'junit_.*\.xml', os.path.basename(f.filename))]
-
-    junit_futures = {}
-    for f in junit_paths:
-        junit_futures[gcs_async.read(f)] = f
-
-    for future in junit_futures:
-        junit = future.get_result()
-        if junit is None:
-            continue
-        failures.extend(parse_junit(junit, junit_futures[future]))
-    failures.sort()
-
-    build_log = None
-    if finished and finished.get('result') != 'SUCCESS' and len(failures) == 0:
-        build_log = gcs_async.read(build_dir + '/build-log.txt').get_result()
-        if build_log:
-            build_log = log_parser.digest(build_log.decode('utf8', 'replace'))
-            logging.info('fallback log parser emitted %d lines',
-                         build_log.count('\n'))
-    return started, finished, failures, build_log
 
 
 @memcache_memoize('pr-details://', expires=60 * 3)
@@ -169,56 +88,6 @@ class IndexHandler(RenderingHandler):
         self.render("index.html", {'jobs': DEFAULT_JOBS})
 
 
-class BuildHandler(RenderingHandler):
-    """Show information about a Build and its failing tests."""
-    def get(self, prefix, job, build):
-        self.check_bucket(prefix)
-        job_dir = '/%s/%s/' % (prefix, job)
-        build_dir = job_dir + build
-        details = build_details(build_dir)
-        if not details:
-            logging.warning('unable to load %s', build_dir)
-            self.render('build_404.html', {"build_dir": build_dir})
-            self.response.set_status(404)
-            return
-        started, finished, failures, build_log = details
-
-        if started:
-            commit = started['version'].split('+')[-1]
-        else:
-            commit = None
-
-        pr = None
-        if prefix.startswith(PR_PREFIX):
-            pr = os.path.basename(prefix)
-        self.render('build.html', dict(
-            job_dir=job_dir, build_dir=build_dir, job=job, build=build,
-            commit=commit, started=started, finished=finished,
-            failures=failures, build_log=build_log, pr=pr))
-
-
-class BuildListHandler(RenderingHandler):
-    """Show a list of Builds for a Job."""
-    def get(self, prefix, job):
-        self.check_bucket(prefix)
-        job_dir = '/%s/%s/' % (prefix, job)
-        fstats = gcs_ls(job_dir)
-        fstats.sort(key=lambda f: pad_numbers(f.filename), reverse=True)
-        self.render('build_list.html',
-                    dict(job=job, job_dir=job_dir, fstats=fstats))
-
-
-
-class JobListHandler(RenderingHandler):
-    """Show a list of Jobs in a directory."""
-    def get(self, prefix):
-        self.check_bucket(prefix)
-        jobs_dir = '/%s' % prefix
-        fstats = gcs_ls(jobs_dir)
-        fstats.sort()
-        self.render('job_list.html', dict(jobs_dir=jobs_dir, fstats=fstats))
-
-
 class PRHandler(RenderingHandler):
     """Show a list of test runs for a PR."""
     def get(self, pr):
@@ -264,9 +133,9 @@ class PRBuildLogHandler(webapp2.RequestHandler):
 
 app = webapp2.WSGIApplication([
     (r'/', IndexHandler),
-    (r'/jobs/(.*)$', JobListHandler),
-    (r'/builds/(.*)/([^/]+)/?', BuildListHandler),
-    (r'/build/(.*)/([^/]+)/(\d+)/?', BuildHandler),
+    (r'/jobs/(.*)$', view_build.JobListHandler),
+    (r'/builds/(.*)/([^/]+)/?', view_build.BuildListHandler),
+    (r'/build/(.*)/([^/]+)/(\d+)/?', view_build.BuildHandler),
     (r'/build/(.*)/([^/]+)/(\d+)/nodelog*', view_logs.NodeLogHandler),
     (r'/pr/(\d+)', PRHandler),
     (r'/pr/?', PRDashboard),
