+++ gubernator/view_base_test.py
@@ -1,0 +1,23 @@
+# Copyright 2016 The Kubernetes Authors All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import unittest
+
+import view_base
+
+
+class HelperTest(unittest.TestCase):
+    def test_pad_numbers(self):
+        self.assertEqual(view_base.pad_numbers('a3b45'),
+                         'a' + '0' * 15 + '3b' + '0' * 14 + '45')
